Origin/frontend/localnotifications (#175)
* imported all modules into intellij

* dbhelper

created dbhelper and methods

* added retrofit

tests for testing the values
and interfaces for ms communications

* making use of retrofit

* implemented apis from other ms to be used

* fixed some errors in api declaration

Also implemented some logic for interacting with filenames

* remove unused api

Also fixed a bug

* Updated devices ms

MS can now save deviceId to family account
retrieve list of devices for user or family account
add/remove devices
check ms health with respect to db connection
talk to user ms to validate users

* updated gitignore and added main class to run the device ms

* added ability for the notif ms to talk to device ms

to retrieve list of topics for a user

* refactored name

* deleted some code and removed some unnecessary import

* updated info for app.properties

* new feature for device ms

If a user admin deletes his account, all the information about his devices will also be deleted

* removed application.yml and new features for users ms

user ms will be able to communicate to device and notif ms to share information

* media ms has ability to delete all files for a deleted family account

* user ms can now finally delete accounts properly

can also talk to other ms if need be

* new api to retrieve the family account for the device

* the media ms can now talk to device ms to retrieve fam acc to save file

* fixed error in test

* make sure to provide the list of devices to be deleted to notif ms

* new notif api to delete all topics

* slightly better at dealing with null pointer exceptions for user ms

* forgot try catch there too

* fixing endpoints + shared preferences

Updated login , register and notification endpoints
Created shared preferences for email, token, remember me state and functionality for auto login when token is present

* better way to get the environment variables into the spring app

* proceeding to catch nullpointerexceptions 

and better deal with environment variables for notifications ms

* small improvements

* better micro service health check

* updated device ms to better deal with npr exceptions

also updated unit test to create docker image
also update app.prop

* better app.prop for building image

* java doc

* make sure to wait to delete topic

* Features for User ms

- Added GsonConverter for retrofit
- Fixed urls in the interfaces
- Now saves the new family account

* added parametrized call to prevent errors

* added json converter and removed / in api def

* added gson converter and updated apis for media ms

* updated ms to use gson converter and apis

* when someone needs to delete the account, 

the family account will then be deleted

* fixed a bug where retrofit would not parse properly

also fixed a null pointer when retrieving list of device ids when there are none

* new api and bugs fixing for the device ms

* fixed a bug where json would not work

also fixed account model so they have the same info as with the other ms

* add check to make sure that user requires password to delete his account

* fixed a bug where the password would not be compared accurately

* fixes the json bug

* actually pings the db to test for app health

and implicitly ignore an error

* Refactoring, fixing some null pointer exceptions and removed 

Amazon Web Services errors

* fixes json bug with retrofit in notif ms

* removed no longer useful tests

* fixed api when talking to device

also made sure that cannot publish if topic does not exists

* to ignore Paw files

* fixed lenient error

+++ Back End/users/src/main/java/com/theelite/users/service/UserServiceImpl.java
@@ -1,30 +1,54 @@
 package com.theelite.users.service;
 
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.theelite.users.communication.DeviceService;
+import com.theelite.users.communication.MediaService;
+import com.theelite.users.communication.NotifService;
 import com.theelite.users.dao.UserDao;
 import com.theelite.users.model.*;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
 
+import java.io.IOException;
 import java.security.SecureRandom;
 import java.util.*;
 
 @Service
 public class UserServiceImpl implements UserService {
-    @Autowired
-    private SecureRandom secureRandom;
+    private final SecureRandom secureRandom;
+    private final Base64.Encoder base64Encoder;
+    private final PasswordEncoder passwordEncoder;
+    private final UserDao userDao;
 
-    @Autowired
-    private Base64.Encoder base64Encoder;
+    private DeviceService deviceService;
+    private NotifService notifService;
+    private MediaService mediaService;
 
-    @Autowired
-    private PasswordEncoder passwordEncoder;
+    public UserServiceImpl(SecureRandom secureRandom, Base64.Encoder base64Encoder, PasswordEncoder passwordEncoder, UserDao userDao, Environment environment) {
+        this.secureRandom = secureRandom;
+        this.base64Encoder = base64Encoder;
+        this.passwordEncoder = passwordEncoder;
+        this.userDao = userDao;
+        this.deviceService = this.buildRetrofitObject(environment.getProperty("device.url"), DeviceService.class);
+        this.notifService = this.buildRetrofitObject(environment.getProperty("notif.url"), NotifService.class);
+        this.mediaService = this.buildRetrofitObject(environment.getProperty("media.url"), MediaService.class);
+    }
 
-    @Autowired
-    private UserDao userDao;
-
+    private void userCreatedNewFamilyAccount(String accountId) {
+        try {
+            notifService.createNewConsumerGroup(accountId).execute();
+        } catch (IOException e) {
+            System.out.println(e.getMessage());
+        } catch (NullPointerException e) {
+            System.out.println("notifService is null; " + e.getMessage());
+        }
+    }
 
     // User Signs up
     @Override
@@ -41,8 +65,13 @@ public class UserServiceImpl implements UserService {
             user.setRole(UserRole.Regular);
         } else {
             // Only user in his family account
-            user.setAccountId(UUID.randomUUID());
             user.setRole(UserRole.Admin);
+            Account newFamAcc = new Account();
+            //Saves the new Family Account
+            userDao.saveNewFamilyAccount(newFamAcc);
+            user.setAccountId(newFamAcc.getAccountId());
+            // to kafka to create new consumer group
+            userCreatedNewFamilyAccount(user.getAccountId().toString());
         }
 
         user.setTokens(Collections.singletonList(userToken));
@@ -67,6 +96,20 @@ public class UserServiceImpl implements UserService {
     @Override
     public boolean deleteUser(User user) {
         if (!userDao.userExistsWithEmail(user.getEmail())) return false;
+        User userInfo = userDao.findById(user.getEmail()).get();
+        if (!passwordEncoder.matches(user.getPassword(), userInfo.getPassword())) return false;
+        String famAcc = userInfo.getAccountId().toString();
+
+        if (userInfo.getRole().equals(UserRole.Admin) && userDao.numberOfAdminsInFamilyAccount(UUID.fromString(famAcc)) == 1) {
+            try {
+                deviceService.familyAccountDeleted(famAcc).execute();
+                notifService.deleteConsumerGroup(famAcc).execute();
+                mediaService.deleteAllForFamilyAccount(famAcc).execute();
+                userDao.deleteFamilyAccount(userInfo.getAccountId());
+            } catch (IOException | NullPointerException e) {
+                System.out.println(e.getMessage());
+            }
+        }
         userDao.deleteById(user.getEmail());
         return true;
     }
@@ -135,6 +178,12 @@ public class UserServiceImpl implements UserService {
     public ResponseEntity<String> getHealth() {
         try {
             userDao.testDatabaseConnection();
+            if (notifService == null)
+                return new ResponseEntity<>("notifService is null", HttpStatus.INTERNAL_SERVER_ERROR);
+            else if (deviceService == null)
+                return new ResponseEntity<>("deviceService is null", HttpStatus.INTERNAL_SERVER_ERROR);
+            else if (mediaService == null)
+                return new ResponseEntity<>("mediaService is null", HttpStatus.INTERNAL_SERVER_ERROR);
         } catch (Exception e) {
             return new ResponseEntity<>("Error with the db somehow", HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -155,5 +204,17 @@ public class UserServiceImpl implements UserService {
 
     public String testToken() {
         return generateNewToken();
+    }
+
+    private <T> T buildRetrofitObject(String url, Class<T> retroClass) {
+        if (url == null || url.isBlank() || url.isEmpty()) {
+            System.out.println("Could not create Retrofit object as url for " + retroClass.getName() + " is null.");
+            return null;
+        }
+        Gson gson = new GsonBuilder()
+                .setLenient()
+                .create();
+        Retrofit retrofit = new Retrofit.Builder().baseUrl(url).addConverterFactory(GsonConverterFactory.create(gson)).build();
+        return retrofit.create(retroClass);
     }
 }
